Frontend. Save the general search dialog window as a separate component (#30) - Improved design - Debouncing starts from the second character of the query
